refactor(vector): move query/result/ordering/record types into vtypes
